v2020.d.2 | Added Notification Sound and Updated Report
----

Version: 2020.d.2
Date: April 28, 2020	18:21

----

A notification sound is added to the application. When the current countdown
is completed, the system will play this sound automatically. The soundOn
variable, which can be toggled by the Sound button, determines whether the
system will play the sound or not.

----

The project report has been updated.